requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
python-dotenv
Pillow
djangorestframework-simplejwt
django-filter
